Rename lsapi.py to mozscape.py.
