+++ python/setup.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 
 # Copyright (c) 2011-2015 SEOmoz, Moz
-# 
+#
 # Permission is hereby granted, free of charge, to any person obtaining
 # a copy of this software and associated documentation files (the
 # "Software"), to deal in the Software without restriction, including
@@ -9,10 +9,10 @@
 # distribute, sublicense, and/or sell copies of the Software, and to
 # permit persons to whom the Software is furnished to do so, subject to
 # the following conditions:
-# 
+#
 # The above copyright notice and this permission notice shall be
 # included in all copies or substantial portions of the Software.
-# 
+#
 # THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
 # EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
 # MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
@@ -27,12 +27,12 @@ except ImportError:
 	from distutils.core import setup
 
 setup(
-	name             = 'lsapi',
-	version          = '0.1.0',
+	name             = 'mozscape',
+	version          = '0.1.1',
 	description      = 'Mozscape API Bindings for Python',
-	author_email     = 'dan@seomoz.org',
+	author_email     = 'dan@moz.com',
 	url              = 'http://github.com/seomoz/SEOmozAPISamples',
-	py_modules       = ['lsapi'],
+	py_modules       = ['mozscape'],
 	license          = 'MIT',
 	platforms        = 'Posix; MacOS X',
 	classifiers      = [
